Move components to their own package
