+++ StaffMod/src/main/kotlin/opekope2/avm_staff/api/item/renderer/StaffRenderer.kt
@@ -27,7 +27,7 @@ import net.minecraft.client.render.model.json.ModelTransformationMode
 import net.minecraft.client.util.math.MatrixStack
 import net.minecraft.item.ItemStack
 import net.minecraft.registry.Registries
-import opekope2.avm_staff.api.staff.StaffRendererPartComponent
+import opekope2.avm_staff.api.component.StaffRendererPartComponent
 import opekope2.avm_staff.api.staffRendererOverrideComponentType
 import opekope2.avm_staff.api.staffRendererPartComponentType
 import opekope2.avm_staff.util.itemStackInStaff
